+++ pkg/defaults/defaults.go
@@ -126,9 +126,3 @@ func HelmChartSubDir() string {
 func PathToHelmChart(name string, version string) string {
 	return def().PathToHelmChart(name, version)
 }
-
-// IsUpgrade determines if we are upgrading a cluster judging by the existence
-// or not of a kubeconfig file in the configuration directory.
-func IsUpgrade() bool {
-	return def().IsUpgrade()
-}
